[CW-000] Changer le int du cash et des points du student pour un bigint

+++ codewars_script.sql
@@ -81,8 +81,8 @@ create table if not exists "user" (
 create table if not exists student (
     da int,
     team_id int,
-    cash int,
-    points int,
+    cash bigint,
+    points bigint,
     constraint fk_team_id foreign key (team_id) references team(id),
     constraint fk_da_student foreign key (da) references "user"(da),
     constraint pk_da_student primary key (da)
